// src/features/Conveyor/interfaces/Conveyor.interface.ts
interface ConveyorInterface {
  x: number; // The x-coordinate of the start Conveyor
  y: number; // The y-coordinate of the start Conveyor
  width: number; // The width of the  Conveyor
  height: number; // The height of the Conveyor
  color: string; // The color of the Conveyor
}

export default ConveyorInterface;